requirements.txt
pipreqs==0.5.0
pytest==8.3.3
python-dotenv==1.0.1
